+++ src/themes/DefaultTheme.js
@@ -1,43 +1,6 @@
-function LightenDarkenColor(color, value) {
-  let usePound = false;
+import { lightenDarkenColor, convertToRgb } from './helpers';
 
-  if (color[0] === "#") {
-    color = color.slice(1);
-    usePound = true;
-  }
-
-  const num = parseInt(color, 16);
-
-  // Red
-  let red = (num >> 16) + value;
-  if (red > 255) red = 255;
-  else if (red < 0) red = 0;
-
-  // Green
-  let green = (num & 0x0000FF) + value;
-  if (green > 255) green = 255;
-  else if (green < 0) green = 0;
-
-  // Blue
-  let blue = ((num >> 8) & 0x00FF) + value;
-  if (blue > 255) blue = 255;
-  else if (blue < 0) blue = 0;
-
-  return (usePound?"#":"") + (green | (blue << 8) | (red << 16)).toString(16);
-}
-
-function convertToRgb(hexVal) {
-  let color = hexVal;
-
-  if (color[0] === "#") color = color.slice(1);
-  const num = parseInt(color, 16);
-  const rgbVal = {
-    red: num >> 16,
-    green: (num >> 8) & 0x00FF,
-    blue: num & 0x0000FF,
-  }
-  return `${rgbVal.red}, ${rgbVal.green}, ${rgbVal.blue}`;
-}
+console.log(lightenDarkenColor);
 
 const colors = {
   purple: '#7A50C7',
@@ -128,7 +91,7 @@ export default {
       backgroundColorActive: gradients.brightActive,
       backgroundColorDisabled: theme.colors.background,
       borderColor: theme.colors.uiLight,
-      borderColorHover: theme.colors.text,
+      borderColorHover: lightenDarkenColor(theme.colors.uiLight, -10),
       textColor: theme.colors.textDark,
       textColorDisabled: theme.colors.uiLight,
       iconColor: theme.colors.textDark,
@@ -161,7 +124,7 @@ export default {
     item: {
       titleColor: theme.colors.uiDark,
       hoverColor: theme.colors.background,
-      activeColor: LightenDarkenColor(theme.colors.background, -2),
+      activeColor: lightenDarkenColor(theme.colors.background, -2),
       iconColor: theme.colors.textDark,
     },
   },
